refactor: replace extension
